added profile link
Wondering why everyone is mixing up the <h3> and <li> tag's in HTML, you need to close the tags following the order hierarchy 
Wouldn't we be closing the tags like this?   <h3> <li> </li> </h3> 

If you just want to increase the font size of the list elements I think it makes sense to create a CSS class and use that instead of using an <h3> tag, 
Since it is bad practice. This can cause accessibility issues to arise when someone is using a screen reader software. As screen reading software organizes and reads
Your text based on the Heading tag hierarchy. 

Here is an example of what I mean. 
<li class="font-large">My name is Rehman, and my profile link is ... </li> 

I've created the CSS class and submitted a pull request with these suggested changes. 
I hope someone else can follow through as well.